Revert #482: "aes: implement VAES AVX and AVX512 backends"
This reverts commit ad83428217e41b33b34c8bfd55424261293e3232.

This implementation uses assembly, but the relevant intrinsics will be
stable in Rust 1.89, and we have a PR open to use them: #491

For a cleaner history, this reverts the assembly implementation so the
intrinsics-based implementation can be cleanly applied to an ASM-free
codebase, rather than as a replacement for the ASM.
