+++ aes/Cargo.toml
@@ -31,7 +31,7 @@ hazmat = [] # Expose cryptographically hazardous APIs
 
 [lints.rust.unexpected_cfgs]
 level = "warn"
-check-cfg = ["cfg(aes_compact)", "cfg(aes_force_soft)", "cfg(aes_avx256_disable)", "cfg(aes_avx512_disable)"]
+check-cfg = ["cfg(aes_compact)", "cfg(aes_force_soft)"]
 
 [package.metadata.docs.rs]
 all-features = true
