+++ aes/src/autodetect.rs
@@ -12,10 +12,12 @@ use core::fmt;
 use core::mem::ManuallyDrop;
 
 #[cfg(target_arch = "aarch64")]
-use crate::armv8 as arch;
+use crate::armv8 as intrinsics;
 
 #[cfg(any(target_arch = "x86_64", target_arch = "x86"))]
-use crate::x86 as arch;
+use crate::ni as intrinsics;
+
+cpufeatures::new!(aes_intrinsics, "aes");
 
 macro_rules! define_aes_impl {
     (
@@ -27,21 +29,21 @@ macro_rules! define_aes_impl {
         doc = $doc:expr,
     ) => {
         mod $module {
-            use super::{arch, soft};
+            use super::{intrinsics, soft};
             use core::mem::ManuallyDrop;
 
             pub(super) union Inner {
-                pub(super) arch: ManuallyDrop<arch::$name>,
+                pub(super) intrinsics: ManuallyDrop<intrinsics::$name>,
                 pub(super) soft: ManuallyDrop<soft::$name>,
             }
 
             pub(super) union InnerEnc {
-                pub(super) arch: ManuallyDrop<arch::$name_enc>,
+                pub(super) intrinsics: ManuallyDrop<intrinsics::$name_enc>,
                 pub(super) soft: ManuallyDrop<soft::$name_enc>,
             }
 
             pub(super) union InnerDec {
-                pub(super) arch: ManuallyDrop<arch::$name_dec>,
+                pub(super) intrinsics: ManuallyDrop<intrinsics::$name_dec>,
                 pub(super) soft: ManuallyDrop<soft::$name_dec>,
             }
         }
@@ -50,7 +52,7 @@ macro_rules! define_aes_impl {
         #[doc = "block cipher"]
         pub struct $name {
             inner: $module::Inner,
-            token: arch::features::aes::InitToken,
+            token: aes_intrinsics::InitToken,
         }
 
         impl KeySizeUser for $name {
@@ -68,7 +70,9 @@ macro_rules! define_aes_impl {
                 use core::ops::Deref;
                 let inner = if enc.token.get() {
                     $module::Inner {
-                        arch: ManuallyDrop::new(unsafe { enc.inner.arch.deref().into() }),
+                        intrinsics: ManuallyDrop::new(unsafe {
+                            enc.inner.intrinsics.deref().into()
+                        }),
                     }
                 } else {
                     $module::Inner {
@@ -86,11 +90,11 @@ macro_rules! define_aes_impl {
         impl KeyInit for $name {
             #[inline]
             fn new(key: &Key<Self>) -> Self {
-                let (token, aes_features) = arch::features::aes::init_get();
+                let (token, aesni_present) = aes_intrinsics::init_get();
 
-                let inner = if aes_features {
+                let inner = if aesni_present {
                     $module::Inner {
-                        arch: ManuallyDrop::new(arch::$name::new(key)),
+                        intrinsics: ManuallyDrop::new(intrinsics::$name::new(key)),
                     }
                 } else {
                     $module::Inner {
@@ -111,7 +115,7 @@ macro_rules! define_aes_impl {
             fn clone(&self) -> Self {
                 let inner = if self.token.get() {
                     $module::Inner {
-                        arch: unsafe { self.inner.arch.clone() },
+                        intrinsics: unsafe { self.inner.intrinsics.clone() },
                     }
                 } else {
                     $module::Inner {
@@ -132,20 +136,38 @@ macro_rules! define_aes_impl {
 
         impl BlockCipherEncrypt for $name {
             fn encrypt_with_backend(&self, f: impl BlockCipherEncClosure<BlockSize = U16>) {
-                if self.token.get() {
-                    unsafe { &self.inner.arch }.encrypt_with_backend(f)
-                } else {
-                    unsafe { &self.inner.soft }.encrypt_with_backend(f)
+                unsafe {
+                    if self.token.get() {
+                        #[target_feature(enable = "aes")]
+                        unsafe fn inner(
+                            state: &intrinsics::$name,
+                            f: impl BlockCipherEncClosure<BlockSize = U16>,
+                        ) {
+                            f.call(state.get_enc_backend());
+                        }
+                        inner(&self.inner.intrinsics, f);
+                    } else {
+                        f.call(&self.inner.soft.get_enc_backend());
+                    }
                 }
             }
         }
 
         impl BlockCipherDecrypt for $name {
             fn decrypt_with_backend(&self, f: impl BlockCipherDecClosure<BlockSize = U16>) {
-                if self.token.get() {
-                    unsafe { &self.inner.arch }.decrypt_with_backend(f)
-                } else {
-                    unsafe { &self.inner.soft }.decrypt_with_backend(f)
+                unsafe {
+                    if self.token.get() {
+                        #[target_feature(enable = "aes")]
+                        unsafe fn inner(
+                            state: &intrinsics::$name,
+                            f: impl BlockCipherDecClosure<BlockSize = U16>,
+                        ) {
+                            f.call(state.get_dec_backend());
+                        }
+                        inner(&self.inner.intrinsics, f);
+                    } else {
+                        f.call(&self.inner.soft.get_dec_backend());
+                    }
                 }
             }
         }
@@ -166,7 +188,7 @@ macro_rules! define_aes_impl {
             #[inline]
             fn drop(&mut self) {
                 if self.token.get() {
-                    unsafe { ManuallyDrop::drop(&mut self.inner.arch) };
+                    unsafe { ManuallyDrop::drop(&mut self.inner.intrinsics) };
                 } else {
                     unsafe { ManuallyDrop::drop(&mut self.inner.soft) };
                 };
@@ -180,7 +202,7 @@ macro_rules! define_aes_impl {
         #[doc = "block cipher (encrypt-only)"]
         pub struct $name_enc {
             inner: $module::InnerEnc,
-            token: arch::features::aes::InitToken,
+            token: aes_intrinsics::InitToken,
         }
 
         impl KeySizeUser for $name_enc {
@@ -190,11 +212,11 @@ macro_rules! define_aes_impl {
         impl KeyInit for $name_enc {
             #[inline]
             fn new(key: &Key<Self>) -> Self {
-                let (token, aes_features) = arch::features::aes::init_get();
+                let (token, aesni_present) = aes_intrinsics::init_get();
 
-                let inner = if aes_features {
+                let inner = if aesni_present {
                     $module::InnerEnc {
-                        arch: ManuallyDrop::new(arch::$name_enc::new(key)),
+                        intrinsics: ManuallyDrop::new(intrinsics::$name_enc::new(key)),
                     }
                 } else {
                     $module::InnerEnc {
@@ -215,7 +237,7 @@ macro_rules! define_aes_impl {
             fn clone(&self) -> Self {
                 let inner = if self.token.get() {
                     $module::InnerEnc {
-                        arch: unsafe { self.inner.arch.clone() },
+                        intrinsics: unsafe { self.inner.intrinsics.clone() },
                     }
                 } else {
                     $module::InnerEnc {
@@ -236,10 +258,19 @@ macro_rules! define_aes_impl {
 
         impl BlockCipherEncrypt for $name_enc {
             fn encrypt_with_backend(&self, f: impl BlockCipherEncClosure<BlockSize = U16>) {
-                if self.token.get() {
-                    unsafe { &self.inner.arch }.encrypt_with_backend(f)
-                } else {
-                    unsafe { &self.inner.soft }.encrypt_with_backend(f)
+                unsafe {
+                    if self.token.get() {
+                        #[target_feature(enable = "aes")]
+                        unsafe fn inner(
+                            state: &intrinsics::$name_enc,
+                            f: impl BlockCipherEncClosure<BlockSize = U16>,
+                        ) {
+                            f.call(state.get_enc_backend());
+                        }
+                        inner(&self.inner.intrinsics, f);
+                    } else {
+                        f.call(&self.inner.soft.get_enc_backend());
+                    }
                 }
             }
         }
@@ -260,7 +291,7 @@ macro_rules! define_aes_impl {
             #[inline]
             fn drop(&mut self) {
                 if self.token.get() {
-                    unsafe { ManuallyDrop::drop(&mut self.inner.arch) };
+                    unsafe { ManuallyDrop::drop(&mut self.inner.intrinsics) };
                 } else {
                     unsafe { ManuallyDrop::drop(&mut self.inner.soft) };
                 };
@@ -274,7 +305,7 @@ macro_rules! define_aes_impl {
         #[doc = "block cipher (decrypt-only)"]
         pub struct $name_dec {
             inner: $module::InnerDec,
-            token: arch::features::aes::InitToken,
+            token: aes_intrinsics::InitToken,
         }
 
         impl KeySizeUser for $name_dec {
@@ -293,7 +324,9 @@ macro_rules! define_aes_impl {
                 use core::ops::Deref;
                 let inner = if enc.token.get() {
                     $module::InnerDec {
-                        arch: ManuallyDrop::new(unsafe { enc.inner.arch.deref().into() }),
+                        intrinsics: ManuallyDrop::new(unsafe {
+                            enc.inner.intrinsics.deref().into()
+                        }),
                     }
                 } else {
                     $module::InnerDec {
@@ -311,11 +344,11 @@ macro_rules! define_aes_impl {
         impl KeyInit for $name_dec {
             #[inline]
             fn new(key: &Key<Self>) -> Self {
-                let (token, aes_features) = arch::features::aes::init_get();
+                let (token, aesni_present) = aes_intrinsics::init_get();
 
-                let inner = if aes_features {
+                let inner = if aesni_present {
                     $module::InnerDec {
-                        arch: ManuallyDrop::new(arch::$name_dec::new(key)),
+                        intrinsics: ManuallyDrop::new(intrinsics::$name_dec::new(key)),
                     }
                 } else {
                     $module::InnerDec {
@@ -336,7 +369,7 @@ macro_rules! define_aes_impl {
             fn clone(&self) -> Self {
                 let inner = if self.token.get() {
                     $module::InnerDec {
-                        arch: unsafe { self.inner.arch.clone() },
+                        intrinsics: unsafe { self.inner.intrinsics.clone() },
                     }
                 } else {
                     $module::InnerDec {
@@ -357,10 +390,19 @@ macro_rules! define_aes_impl {
 
         impl BlockCipherDecrypt for $name_dec {
             fn decrypt_with_backend(&self, f: impl BlockCipherDecClosure<BlockSize = U16>) {
-                if self.token.get() {
-                    unsafe { &self.inner.arch }.decrypt_with_backend(f)
-                } else {
-                    unsafe { &self.inner.soft }.decrypt_with_backend(f)
+                unsafe {
+                    if self.token.get() {
+                        #[target_feature(enable = "aes")]
+                        unsafe fn inner(
+                            state: &intrinsics::$name_dec,
+                            f: impl BlockCipherDecClosure<BlockSize = U16>,
+                        ) {
+                            f.call(state.get_dec_backend());
+                        }
+                        inner(&self.inner.intrinsics, f);
+                    } else {
+                        f.call(&self.inner.soft.get_dec_backend());
+                    }
                 }
             }
         }
@@ -381,7 +423,7 @@ macro_rules! define_aes_impl {
             #[inline]
             fn drop(&mut self) {
                 if self.token.get() {
-                    unsafe { ManuallyDrop::drop(&mut self.inner.arch) };
+                    unsafe { ManuallyDrop::drop(&mut self.inner.intrinsics) };
                 } else {
                     unsafe { ManuallyDrop::drop(&mut self.inner.soft) };
                 };
