+++ aes/src/hazmat.rs
@@ -21,7 +21,7 @@ pub type Block8 = cipher::array::Array<Block, cipher::consts::U8>;
 use crate::armv8::hazmat as intrinsics;
 
 #[cfg(all(any(target_arch = "x86_64", target_arch = "x86"), not(aes_force_soft)))]
-use crate::x86::ni::hazmat as intrinsics;
+use crate::ni::hazmat as intrinsics;
 
 #[cfg(all(
     any(target_arch = "x86", target_arch = "x86_64", target_arch = "aarch64"),
