--- aes/src/x86/vaes512.rs
@@ -1,3 +1,0 @@
-pub(super) mod aes128;
-pub(super) mod aes192;
-pub(super) mod aes256;
